backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
scikit-learn==1.3.2
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
huggingface-hub==0.19.0
diffusers==0.24.0
transformers==4.35.0
accelerate==0.24.0
matplotlib==3.8.2
seaborn==0.13.0
numpy==1.24.3
